requirements.txt

---

# 2. `requirements.txt`

```txt
torch>=2.3.0
torchvision>=0.18.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.29.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
boto3>=1.34.0
locust>=2.18.0
pydantic>=2.5.0
python-multipart>=0.0.6
seaborn>=0.13.0
